Resolved #2
I stored the MDL that used to lock Pages at SNIFF_OBJECT.
But the SNIFF_OBJECT had been freed in the each LockAndMapMemory Routine.
So It had failed to unmap & unlock in the UnMapAndUnLockMemory Routine.

I change the location of the MDl from the SNIFF_OBJECT to the TARGET_OBJECT.